Moved hyperbolic tangent activation to it's own file.
