+++ src/activation/mod.rs
@@ -2,9 +2,11 @@ use std::f64;
 
 pub mod sigmoid;
 pub mod identity;
+pub mod hyperbolictangent;
 
 pub use self::sigmoid::Sigmoid;
 pub use self::identity::Identity;
+pub use self::hyperbolictangent::HyperbolicTangent;
 
 /// Activation functions
 pub trait Activation {
@@ -13,25 +15,6 @@ pub trait Activation {
     fn calc(&self, x: f64) -> f64;
     // Derivative
     fn derivative(&self, x: f64) -> f64;
-}
-
-pub struct HyperbolicTangent;
-
-impl Activation for HyperbolicTangent {
-    fn new() -> HyperbolicTangent {
-        return HyperbolicTangent;
-    }
-
-    /// Calculates the tanh of input `x`
-    fn calc(&self, x: f64) -> f64 {
-        x.tanh()
-    }
-
-    /// Calculates the Derivative tanh of input `x`
-    fn derivative(&self, x: f64) -> f64 {
-        let tanh_factor = x.tanh();
-        1f64 - (tanh_factor * tanh_factor)
-    }
 }
 
 pub struct SoftPlus;
@@ -81,21 +64,8 @@ impl Activation for RectifiedLinearUnit {
 #[cfg(test)]
 mod tests {
     use super::Activation;
-    use super::HyperbolicTangent;
     use super::SoftPlus;
     use super::RectifiedLinearUnit;
-
-    #[test]
-    fn tanh_test() {
-        let activation = HyperbolicTangent::new();
-        assert_approx_eq!(activation.calc(3f64), 0.995054754f64);
-    }
-
-    #[test]
-    fn tanh_derivative_test() {
-        let activation = HyperbolicTangent::new();
-        assert_approx_eq!(activation.derivative(3f64), 0.0098660372f64);
-    }
 
     #[test]
     fn softplus_test() {
